refactor(cache): split cache.js into focused modules under cache/
lib/interceptor/cache.js (1390 lines) and cache-revalidation.js move into
lib/interceptor/cache/ as eight focused modules:

- index.js — interceptor factory: read-path decision logic, origins gate
- cache-handler.js — CacheHandler (write path, storability gates)
- revalidation.js — RevalidationHandler + backgroundRefresh
- invalidation-handler.js — InvalidationHandler (unsafe methods, RFC 9111 §4.4)
- serve.js — serveFromCache + the undici:cache lookup-doc emitter
- freshness.js — pure RFC 9111 lifetime/age/entry-times math
- headers.js — parseVary, conditionalHeaders, weakMatch, isEtagUsable
- store.js — makeKey, getStore, tryGetEntry, cacheOptsOf

All code moved verbatim; only imports and stale comment cross-references
changed. The deliberate index.js <-> cache-revalidation.js import cycle
(and its TDZ live-binding invariant) is gone: revalidation.js now imports
only leaf modules, so the dependency graph is acyclic.

Test imports that reached into the old paths are updated to the new
concrete modules; stale path mentions in comments/README updated.

Verified: eslint + prettier clean; all cache test files pass in
isolation; cache-tests conformance suite at the zero-required-failures
baseline.
