+++ lib/index.js
@@ -17,7 +17,7 @@ export const interceptors = {
   log: (await import('./interceptor/log.js')).default,
   redirect: (await import('./interceptor/redirect.js')).default,
   proxy: (await import('./interceptor/proxy.js')).default,
-  cache: (await import('./interceptor/cache.js')).default,
+  cache: (await import('./interceptor/cache/index.js')).default,
   requestId: (await import('./interceptor/request-id.js')).default,
   dns: (await import('./interceptor/dns.js')).default,
   lookup: (await import('./interceptor/lookup.js')).default,
