+++ lib/interceptor/cache/revalidation.js
@@ -1,17 +1,15 @@
-import { parseCacheControl } from '../utils.js'
-import { isHopByHop } from './proxy.js'
+import { parseCacheControl } from '../../utils.js'
+import { isHopByHop } from '../proxy.js'
+import { CacheHandler } from './cache-handler.js'
 import {
-  CacheHandler,
   DEFAULT_MAX_ENTRY_TTL,
-  cacheOptsOf,
   computeEntryTimes,
-  conditionalHeaders,
   determineAge,
   determineLifetime,
-  isEtagUsable,
-  parseVary,
-  serveFromCache,
-} from './cache.js'
+} from './freshness.js'
+import { conditionalHeaders, isEtagUsable, parseVary } from './headers.js'
+import { serveFromCache } from './serve.js'
+import { cacheOptsOf } from './store.js'
 
 const NOOP = () => {}
 
@@ -79,7 +77,7 @@ export class RevalidationHandler {
   #lookupMs
   /** @type {null | 'pass' | 'validated' | 'stale'} */
   #mode = null
-  /** @type {CacheHandler | import('../utils.js').DecoratorHandler | object | null} */
+  /** @type {CacheHandler | import('../../utils.js').DecoratorHandler | object | null} */
   #inner = null
   /** @type {((reason?: any) => void) | null} */
   #abort = null
